PHP 8 Fixes #158

+++ inc/classes/ChWsbAdminTools.php
@@ -441,12 +441,12 @@ EOF;
     function performInstalCheck() { //check requirements
         $aErrors = array();
 
-        $aErrors[] = (ini_get('register_globals') == 0) ? '' : '<font color="red">register_globals is On (warning, you should have this param in the Off state, or your site will be unsafe)</font>';
-        $aErrors[] = (ini_get('safe_mode') == 0) ? '' : '<font color="red">safe_mode is On, disable it</font>';
+        $aErrors[] = ((int)ini_get('register_globals') == 0) ? '' : '<font color="red">register_globals is On (warning, you should have this param in the Off state, or your site will be unsafe)</font>';
+        $aErrors[] = ((int)ini_get('safe_mode') == 0) ? '' : '<font color="red">safe_mode is On, disable it</font>';
         $aErrors[] = (((int)phpversion()) < 4) ? '<font color="red">PHP version too old, update server please</font>' : '';
         $aErrors[] = (!extension_loaded( 'mbstring')) ? '<font color="red">mbstring extension not installed. <b>Warning!</b> Cheetah cannot work without <b>mbstring</b> extension.</font>' : '';
-        $aErrors[] = (ini_get('short_open_tag') == 0 && version_compare(phpversion(), "5.4", "<") == 1) ? '<font color="red">short_open_tag is Off (must be On!)<b>Warning!</b> Cheetah cannot work without <b>short_open_tag</b>.</font>' : '';
-        $aErrors[] = (ini_get('allow_url_include') == 0) ? '' : '<font color="red">allow_url_include is On (warning, you should have this param in the Off state, or your site will be unsafe)</font>';
+        $aErrors[] = ((int)ini_get('short_open_tag') == 0 && version_compare(phpversion(), "5.4", "<") == 1) ? '<font color="red">short_open_tag is Off (must be On!)<b>Warning!</b> Cheetah cannot work without <b>short_open_tag</b>.</font>' : '';
+        $aErrors[] = ((int)ini_get('allow_url_include') == 0) ? '' : '<font color="red">allow_url_include is On (warning, you should have this param in the Off state, or your site will be unsafe)</font>';
 
         $aErrors = array_diff($aErrors, array('')); //delete empty
         if (count($aErrors)) {
@@ -569,7 +569,7 @@ EOF;
 
     function isAllowUrlInclude()
     {
-        $sAllowUrlInclude = ini_get('allow_url_include');
+        $sAllowUrlInclude = (int)ini_get('allow_url_include');
         return !($sAllowUrlInclude == 0);
     }
